feat: add static components to reduce flickering
