+++ apps/cli/src/components/shared/input/select.tsx
@@ -34,16 +34,16 @@ export const Select = <T extends string>({
 }: SelectProps<T>) => {
   const [selectedValue, setSelectedValue] = useState('');
   const [isSuccessSubmitted, setIsSuccessSubmitted] = useState(false);
-  const { setRawMode } = useTerminalState();
+  // const { setRawMode } = useTerminalState();
 
-  useEffect(() => {
-    if (showPrompt && !isSuccessSubmitted) {
-      setRawMode(true);
-    }
-    return () => {
-      setRawMode(false);
-    };
-  }, [showPrompt, isSuccessSubmitted, setRawMode]);
+  // useEffect(() => {
+  //   if (showPrompt && !isSuccessSubmitted) {
+  //     // setRawMode(true);
+  //   }
+  //   return () => {
+  //     setRawMode(false);
+  //   };
+  // }, [showPrompt, isSuccessSubmitted, setRawMode]);
 
   if (selectedValue && status === 'success' && !isSuccessSubmitted) {
     setIsSuccessSubmitted(true);
