+++ apps/cli/src/components/shared/input/text-input.tsx
@@ -1,10 +1,9 @@
 import type { UserMessageLimit } from '@appdotbuild/core';
-import { Spinner } from '@inkjs/ui';
+import { TextInput as InkTextInput, Spinner } from '@inkjs/ui';
 import type { MutationStatus } from '@tanstack/react-query';
 import { Box, Text } from 'ink';
 import { useEffect, useState } from 'react';
 import { Panel } from '../display/panel.js';
-import { BufferedTextInput } from './buffered-text-input.js';
 
 export interface TextInputProps {
   question?: string;
@@ -50,43 +49,44 @@ export function TextInput({
   if (!showPrompt) return null;
 
   return (
-    <Panel title={question} variant="default" boxProps={{ width: '100%' }}>
-      <Box flexDirection="column" gap={1}>
-        <Box>
-          {submittedValue ? (
-            <>
-              <Text color="blue">❯ </Text>
+    <Box flexDirection="column">
+      <Panel title={question} variant="default" boxProps={{ width: '100%' }}>
+        <Box flexDirection="column" gap={1}>
+          <Box>
+            <Text color="blue">❯ </Text>
+            {submittedValue ? (
               <Text color="gray">{submittedValue}</Text>
-            </>
-          ) : (
-            <BufferedTextInput
-              placeholder={placeholder}
-              onSubmit={(value) => {
-                setSubmittedValue(value);
-                onSubmit(value);
-              }}
-              isDisabled={
-                userMessageLimit?.isUserLimitReached || status === 'pending'
-              }
-            />
+            ) : (
+              <InkTextInput
+                placeholder={placeholder}
+                onSubmit={(value) => {
+                  setSubmittedValue(value);
+                  onSubmit(value);
+                }}
+                isDisabled={
+                  userMessageLimit?.isUserLimitReached || status === 'pending'
+                }
+                {...textInputProps}
+              />
+            )}
+          </Box>
+          {status === 'pending' && (
+            <Box gap={1}>
+              <Spinner />
+              <Text color="yellow">{loadingText}</Text>
+            </Box>
           )}
         </Box>
-        {status === 'pending' && (
-          <Box gap={1}>
-            <Spinner />
-            <Text color="yellow">{loadingText}</Text>
-          </Box>
-        )}
+      </Panel>
 
-        {userMessageLimit && (
-          <Box justifyContent="flex-end" marginTop={1}>
-            <Text color={!userMessageLimit.isUserLimitReached ? 'gray' : 'red'}>
-              {userMessageLimit.remainingMessages} /{' '}
-              {userMessageLimit.dailyMessageLimit} messages remaining
-            </Text>
-          </Box>
-        )}
-      </Box>
-    </Panel>
+      {userMessageLimit && (
+        <Box justifyContent="flex-end">
+          <Text color={!userMessageLimit.isUserLimitReached ? 'gray' : 'red'}>
+            {userMessageLimit.remainingMessages} /{' '}
+            {userMessageLimit.dailyMessageLimit} messages remaining
+          </Text>
+        </Box>
+      )}
+    </Box>
   );
 }
